購物車功能 - 修改購物車前確認內容是否為該用戶所有

+++ src/main/java/com/justinwu/springbootmall/service/impl/CartServiceImpl.java
@@ -1,11 +1,15 @@
 package com.justinwu.springbootmall.service.impl;
 
 import com.justinwu.springbootmall.dao.CartDao;
+import com.justinwu.springbootmall.dao.UserDao;
 import com.justinwu.springbootmall.dto.CartItemRequest;
 import com.justinwu.springbootmall.model.CartItem;
+import com.justinwu.springbootmall.model.User;
 import com.justinwu.springbootmall.service.CartService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -14,6 +18,9 @@ public class CartServiceImpl implements CartService {
 
     @Autowired
     private CartDao cartDao;
+
+    @Autowired
+    private UserDao userDao;
 
     @Override
     public List<CartItem> getCartByUserId(Integer userId) {
@@ -28,21 +35,63 @@ public class CartServiceImpl implements CartService {
     }
 
     @Override
-    public Integer createCartItem(CartItemRequest cartItemRequest) {
+    public Integer createCartItem(Integer userId, CartItemRequest cartItemRequest) {
+        if(!isUserExist(userId)){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "用戶不存在");
+        }
+
         //新增購物車商品
-        Integer cartItemId = cartDao.createCartItem(cartItemRequest);
+        Integer cartItemId = cartDao.createCartItem(userId, cartItemRequest);
 
         return cartItemId;
     }
 
     @Override
-    public void updateCartItem(Integer cartItemId, CartItemRequest cartItemRequest) {
+    public void updateCartItem(Integer userId, Integer cartItemId, CartItemRequest cartItemRequest) {
+        if(!isUserExist(userId)){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "用戶不存在");
+        }
+        else if(!isBelong(userId,cartItemId)){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "該用戶無此購物車商品");
+        }
+
         //更新購物車商品
-        cartDao.updateCartItem(cartItemId, cartItemRequest);
+        cartDao.updateCartItem(userId, cartItemId, cartItemRequest);
     }
 
     @Override
-    public void deleteCartItemById(Integer cartItemId) {
+    public void deleteCartItemById(Integer userId, Integer cartItemId) {
+        if(!isUserExist(userId)){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "用戶不存在");
+        }
+        else if(!isBelong(userId,cartItemId)){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "該用戶無此購物車商品");
+        }
+
         cartDao.deleteCartItemById(cartItemId);
+    }
+
+    @Override
+    public Boolean isUserExist(Integer userId){
+        //檢查 user 是否存在
+        User user = userDao.getUserById(userId);
+
+        if (user == null){
+            return false;
+        }
+        else{
+            return true;
+        }
+    }
+
+    @Override
+    public Boolean isBelong(Integer userId, Integer cartItemId){
+        CartItem cartItem = cartDao.getCartItemById(cartItemId);
+        if(cartItem.getUserId().equals(userId)){
+            return true;
+        }
+        else{
+            return false;
+        }
     }
 }
